backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
supabase>=2.4.3
python-dotenv==1.0.1
aiohttp==3.9.3
apscheduler==3.10.4
lxml==5.1.0
python-dateutil==2.8.2
httpx>=0.27.0
python-multipart==0.0.9
dramatiq[redis]==1.17.0
redis==5.0.8
prometheus-client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
